Fiks custom feilmelding for tilogmed felt
