+++ src/frontend/components/Felleskomponenter/UtenlandsoppholdModal/UtenlandsoppholdModal.tsx
@@ -2,12 +2,17 @@ import React from 'react';
 
 import { useIntl } from 'react-intl';
 
-import { ESvar, ISODateString } from '@navikt/familie-form-elements';
+import { ESvar } from '@navikt/familie-form-elements';
 
 import { IUtenlandsperiode } from '../../../typer/person';
 import { EUtenlandsoppholdÅrsak } from '../../../typer/utenlandsopphold';
-import { dagensDato, ettÅrTilbakeDato, gårsdagensDato } from '../../../utils/dato';
 import { svarForSpørsmålMedUkjent } from '../../../utils/spørsmål';
+import {
+    harTilhørendeFomFelt,
+    hentFomAvgrensningPåTilDato,
+    hentTomAvgrensningPåFraDato,
+    hentTomAvgrensningPåTilDato,
+} from '../../../utils/utenlandsopphold';
 import Datovelger from '../Datovelger/Datovelger';
 import { LandDropdown } from '../Dropdowns/LandDropdown';
 import StyledDropdown from '../Dropdowns/StyledDropdown';
@@ -36,51 +41,6 @@ interface Props extends ReturnType<typeof useModal>, IUseUtenlandsoppholdSkjemaP
     tilDatoUkjentLabelSpråkId: string;
     onLeggTilUtenlandsperiode: (periode: IUtenlandsperiode) => void;
 }
-
-export const hentTomAvgrensningPåFraDato = (
-    utenlandsoppÅrsak: EUtenlandsoppholdÅrsak | ''
-): ISODateString | undefined => {
-    switch (utenlandsoppÅrsak) {
-        case EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_FRA_NORGE:
-        case EUtenlandsoppholdÅrsak.OPPHOLDER_SEG_UTENFOR_NORGE:
-            return dagensDato();
-        case EUtenlandsoppholdÅrsak.HAR_OPPHOLDT_SEG_UTENFOR_NORGE:
-            return gårsdagensDato();
-        default:
-            return undefined;
-    }
-};
-export const hentFomAvgrensningPåTilDato = (
-    utenlandsoppÅrsak: EUtenlandsoppholdÅrsak | ''
-): ISODateString | undefined => {
-    switch (utenlandsoppÅrsak) {
-        case EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_TIL_NORGE:
-            return ettÅrTilbakeDato();
-        default:
-            return undefined;
-    }
-};
-
-export const hentTomAvgrensningPåTilDato = (
-    utenlandsoppÅrsak: EUtenlandsoppholdÅrsak | ''
-): ISODateString | undefined => {
-    switch (utenlandsoppÅrsak) {
-        case EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_TIL_NORGE:
-        case EUtenlandsoppholdÅrsak.HAR_OPPHOLDT_SEG_UTENFOR_NORGE:
-            return dagensDato();
-        default:
-            return undefined;
-    }
-};
-
-export const harTilhørendeFomFelt = (
-    utenlandsoppholdÅrsak: EUtenlandsoppholdÅrsak | ''
-): boolean => {
-    return (
-        utenlandsoppholdÅrsak === EUtenlandsoppholdÅrsak.HAR_OPPHOLDT_SEG_UTENFOR_NORGE ||
-        utenlandsoppholdÅrsak === EUtenlandsoppholdÅrsak.OPPHOLDER_SEG_UTENFOR_NORGE
-    );
-};
 
 export const UtenlandsoppholdModal: React.FC<Props> = ({
     landFeilmeldingSpråkIds,
